Privatise tick to TickContainer
